split QuestionPackCard into a component for re-usability
